requirements.txt
absl-py==2.1.0
adios2==2.10.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
configmypy==0.1.0
contourpy==1.2.1
cupy-cuda12x==13.2.0
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.14.2
dill==0.3.8
dm-haiku==0.0.12
docker-pycreds==0.4.0
etils==1.5.2
exceptiongroup==1.2.1
executing==2.0.1
fastrlock==0.8.2
filelock==3.14.0
flax==0.8.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.0
h5py==3.11.0
hjson==3.1.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.18.1
jax==0.4.28
jax-cuda12-pjrt==0.4.28
jax-cuda12-plugin==0.4.28
-e ./solvers/JAXFLUIDS
jaxlib==0.4.28
jedi==0.19.1
jinja2==3.1.4
jmp==0.0.4
joblib==1.4.2
jupyter-client==8.6.2
jupyter-core==5.7.2
kiwisolver==1.4.5
lbmpy==1.3.3
-e ./solvers/lbmpy-atmo-dl
lightning==2.2.5
lightning-utilities==0.11.2
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mpi4py==3.1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
neuraloperator==0.3.0
ninja==1.11.1.1
nose==1.3.7
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.5.40
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.15
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.45
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pydantic==2.7.2
pydantic-core==2.18.3
pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
pystencils==1.3.3
pytest==8.2.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.5
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.13.1
sentry-sdk==2.3.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensordict==0.4.0
tensorly==0.8.1
tensorly-torch==0.4.0
tensorstore==0.1.60
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torch-harmonics==0.6.5
torchaudio==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.0
typing-extensions==4.12.0
unique-names-generator==1.0.2
urllib3==1.26.18
wandb==0.17.0
wcwidth==0.2.13
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.0
